refactor(skills): 抽取 SkillDetailModal 为独立组件
